requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.0
    # via django
django==4.0.3
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-filter==21.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pytz==2021.3
    # via djangorestframework
sqlparse==0.4.2
    # via django

# The following packages are considered to be unsafe in a requirements file:
# setuptools
